server/requirements.txt
FlightRadarAPI===1.3.25
requests===2.31.0
geographiclib===1.52
pytz===2024.1
python-dotenv==1.0.1
